pyproject.toml
[tool.poetry]
name = "cybernexvpn"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
drf-spectacular = "^0.28.0"
django-environ = "^0.12.0"
aiohttp = "^3.11.12"
pydantic = "^2.10.6"
python-dateutil = "^2.9.0.post0"
wireguard-tools = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
